requirements.txt
## Requirements
Tensorboard >= 2.14.0
PyYaml >= 6.0.2
einops >=0.8.0
scipy >=1.10.1
numpy>=1.24.1
Pillow>=10.2.0
transformers>=4.45.2
